pyproject.toml
[project]
name = "levo-survey-api"
version = "0.1.0"
description = "Levo Survey is a web application to create and distribute surveys to collect user feedbacks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "flask-migrate>=4.0.0",
    "flask-cors>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "marshmallow>=3.20.0",
    "flask-marshmallow>=0.15.0",
    "marshmallow-sqlalchemy>=0.29.0",
    "email-validator>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "factory-boy>=3.3.0",
    "flask-smorest>=0.46.1",
    "flask-injector>=0.15.0",
]
